replace OptionsListUtilsTest.ts content with migration files content, remove migration file
